update(c4): stop implement succeed steps of c4 compiler
Implementation of c4 take lots of tricky programming skills, but at the
same time, it use tons of global variables and make lexer, parser,
semantic analysis and code generation become tight coupling, which isn't
nice or graceful programming paradigm for writing compiler.

Obviously, source code of c4 is awesome to learn, with a goal of
enhancing programming trick, not with a goal of learning how to write a
decoupled compiler.

issue sabertazimi/awesome-notes#25